// src/app/enums/player-command.enum.ts
export enum PLAYER_COMMAND {
  NO_COMMAND = 'NO COMMAND',
  NOTHING_COMMAND = 'NOTHING',
  TURN_LEFT_COMMAND = 'TURN_LEFT',
  TURN_RIGHT_COMMAND = 'TURN_RIGHT',
  ACCELERATE_COMMAND = 'ACCELERATE',
  DECELERATE_COMMAND = 'DECELERATE',
  USE_BOOST_COMMAND = 'USE_BOOST',
  USE_OIL_COMMAND = 'USE_OIL',
}